Makefile: add rule to bootstrap pip
This is a simple rule to get pip installed on a system with Python
alone.

In theory, most environments with Python will have pip and setuptools.
For instance, in Fedora, those are dependecies of the python package
itself.

But, for completeness, since QEMU depends on nothing besides Python
itself, let's the hability to install and uninstall pip easily.

TODO:
 1) skip installation attempt if pip is already detected
 2) better cleanup, also only acting if pip seems to have
    been installed by us

Signed-off-by: Cleber Rosa <crosa@redhat.com>