scripts/qemu.py: Add arch parameter and automatic arch guessing
The binary parameter is mandatory, and central to QEMUMachine usage.
Some information, such as the intended architecture, can either be
given explicitly, or can be extracted from the binary itself.

This change provides a generic function, qmp_execute(), that executes
a QEMU binary with the sole purpose of retrieving information from
a binary.  This function is then applied to query the target name
from the given binary and use it as the architecture on QEMUMachine.

Given that many tests depend on the current behavior, and for the sake
of clarity, the automatic setting of the archicture is not done by
default, it requires the "automatic_devices=True" to be given as a
parameter to QEMUMachine.

Signed-off-by: Cleber Rosa <crosa@redhat.com>