QMP: read greeting message at connect() time
Based on the QMP spec documentation, and by drawing parallels between
with a human client, the QEMUMonitorProtocol class should consume
the greeting message at connection time.

For instance, the following "low-level" use case should is perfectly
legal and should be possible:

   >>> client = QEMUMonitorProtocol("address")
   >>> client.connect(False)
   >>> print(client.cmd("qmp_capabilities"))
   {"QMP": {"version": {"qemu": {"micro": 50, "minor": 12, "major": 2},
   "package": "v2.12.0-258-gc8b7e627b4"}, "capabilities": []}}

But it's not with the implicit consuming of the greeting right before
the capabilities negotiation phase.

To make for better individual testing of the connect and capabilities
negotiation phase, let's make the now private methods part of the
public API.

Signed-off-by: Cleber Rosa <crosa@redhat.com>