qemu.py: Add arch parameter and probe
The binary parameter is mandatory, and central to QEMUMachine usage.
Some information, such as the intended architecture, can either be
given explicitly, or can be extracted from the binary itself.

This change provides a generic function, qmp_execute(), that executes
a QEMU binary with the sole purpose of retrieving information from a
binary.  This function is then used on the utility
qemu_bin_probe_arch() to query the target name from the given binary.
Finally, it's used on QEMUMachine.probe_arch() which will be used
to setup archictured (and machine model specific) console devices.

Signed-off-by: Cleber Rosa <crosa@redhat.com>