qemu.py: introduce add_console() method
The add_console() method is intended to ease higher level use cases
that require a console device.

The amount of inteligence is limited on purpose, requiring either
the device type explicitly, or the existence of an archicture and
machine (pattern) entry in the CONSOLE_DEV_TYPE dictionary.

Because of the console device type selection criteria (by machine
type), users should also be able to define that.  It'll then be used
for both '-machine' and for the console device type selection.

Users of the add_console() method will certainly be interested in
accessing the console device, and for that a console_address property
has been added.

Signed-off-by: Cleber Rosa <crosa@redhat.com>