QMP: add specific data error handling exception
The ValueError, raised by the JSON library, is an extremely
generalistic exception.  For users of the QMP Python classes, it's
useful to be able to catch the exact nature of an error.

TODO: give more examples of code that can raise ValueError and is not
related to data being received/parsed by the QMPMonitorProtocol class.

Signed-off-by: Cleber Rosa <crosa@redhat.com>