Acceptance Tests: add example test for Linux guest and console usage
The previous patches would be mostly unjustifiable without users.
This brings one minimalist example, inspired by a previous experiment,
in which a Linux kernel is booted, and the console is checked for
the presence of a given kernel generated output.

In this particular example, the test checks for the command line
that was passed, but it could just as well serve the purpose of
device detection verification.

Signed-off-by: Cleber Rosa <crosa@redhat.com>