RFC: scripts/qemu.py: allow access to the list of extra arguments
There's been discussions on what kind of access users of the
QEMUMachine class should be given.  Suggestions included adding
a new API (add_args()), a rename of "protected _args" to "args"
and possibly others.

As a way to revive the discussion, this patch adds a property that
makes "_args" available as a property named "args".

Signed-off-by: Cleber Rosa <crosa@redhat.com>