RFC: Acceptance test: add regression test for memory address rewrite with ranges
Because of the acknowledged limitations described here:

https://trello.com/c/O882Vulm/44-qemumachine-even-if-we-expect-qemu-to-exit-with-an-error

And being worked on, but still not solved, writing tests that expect
the qemu process to not succeed are not good candidates to use
QEMUMachine, and consequently, avocado_qemu.Test.

This made me evaluate, and realize, that a number of valid testing
scenarios may and should rely on less.  It seems that the base
avocado.Test class may be a good candidate for those.

This test relied on the original patch command line example, but added
"-nodefaults -nographic -S", which seem to be good practice for most
testing purposes, and removed the specified machine type.

Also, the test seems applicable to the ppc64 target, but isn't to
aarch64 (due to a missing machine type, and when given one such as
virt, it's not capable of having the pc-dimm devices).

Reference: https://lists.gnu.org/archive/html/qemu-devel/2018-10/msg05155.html
Signed-off-by: Cleber Rosa <crosa@redhat.com>