Acceptance tests: use "arch:" tag to filter target specific tests
Currently, the only test that contains some target architecture
information is "boot_linux.py" which test contains a "x86_64" tag.
But that tag is not respected in the default execution, that is, "make
check-acceptance" doesn't do anything with it.

That said, even the target architecture handling currently present in
the "avocado_qemu.Test" class, class is pretty limited, and to be
honest, broken in a number of ways -- and not at all not at all
related to tags.  For instance, by default, it chooses a target based
on the host architecture.

Because the original implementation of the tags feature in Avocado did
not include any time of namespace or "key:val" mechanism, no tag has
relation to another tag.  This demands that, based on the current
implementation we set tag that can run in any target arch to be tagged
accordingly.

Signed-off-by: Cleber Rosa <crosa@redhat.com>