Makefile.tests: check for SPICE
While we don't have a mechanism that can either transparently forward
build configuration information to tests, or some better runtime
instrospection on capabilities, let's expose if SPICE is enabled on
the build system.

The approach chosen here is to let tests that require a feature, such
as SPICE, to mark it as such.  Then, if the QEMU build was not
configured with SPICE, an exclusion rule is applied to those tests so
that they're not even attempted to be executed.

This also bumps the Avocado version to one that includes a fix to the
parsing of multiple and mix "key:val" and simple tag values.

Signed-off-by: Cleber Rosa <crosa@redhat.com>