arm/aarch64: reproducer for triaging possible bug
After booting a Linux kernel on both arm and aarch64, the QMP sockets
gets unresponsive.  Initially, this was thought to be limited to
"quit" commands, but it reproduced with others (such as in this
reproducer).  This is a partial log output:

   >>> {'execute': 'qmp_capabilities'}
   <<< {'return': {}}
   Booting Linux on physical CPU 0x0000000000 [0x410fd034]
   Linux version 4.18.16-300.fc29.aarch64 (mockbuild@buildvm-aarch64-02.arm.fedoraproject.org) (gcc version 8.2.1 20180801 (Red Hat 8.2.1-2) (GCC)) #1 SMP Sat Oct 20 23:12:22 UTC 2018
   ...
   Policy zone: DMA32
   Kernel command line: printk.time=0 console=ttyAMA0
   >>> {'execute': 'stop'}
   <<< {'timestamp': {'seconds': 1558370331, 'microseconds': 470173}, 'event': 'STOP'}
   <<< {'return': {}}
   >>> {'execute': 'cont'}
   <<< {'timestamp': {'seconds': 1558370331, 'microseconds': 470849}, 'event': 'RESUME'}
   <<< {'return': {}}
   >>> {'execute': 'stop'}

Sometimes it takes just the first "stop" command.  Overall, I was able
to reproduce 100% of times when applied on top of
6d8e75d41c58892ccc5d4ad61c4da476684c1c83.

A "bug:triage" tag was added for the sake of making the
listing/execution of those simpler.  The command line one can use to
reproduce this:

   $ make check-venv
   $ ./tests/venv/bin/avocado <run|list> -t bug:triage tests/acceptance

Or to look at the test results as they're produced:

   $ ./tests/venv/bin/avocado --show=test run -t bug:triage tests/acceptance

Signed-off-by: Cleber Rosa <crosa@redhat.com>