S390: protect test against guest panic and QEMU shutdown
Because of S390's support for guest panic events, and the fact
that this test boots a Linux kernel without an initrd and/or
root filesystem, it will eventually panic.

Depending on the amount of time the test sticks around after finding
the expected kernel output, it may be see guest panic, have QEMU
shutdown and a test error condition.

A reproducer for the situation this protects against is to add
the following to the end of the test code:

    import time
    for _ in range(10):
        self.vm.qmp('query-status')
        time.sleep(1)

Signed-off-by: Cleber Rosa <crosa@redhat.com>