[RFC] Add "boot_linux" test for s390x and s390-ccw-virtio machine type
This is not meant for inclusion at this point, given that the Fedora
30 Cloud image for s390x was never put at the "right" location in the
Fedora servers due to a Release Engineering mistake.  Thus, it
requires a hack on the Avocado vmimage library, which is fulfilled
here by a change in the requirements.txt file.

Still, this should work as intended, and early comments and testing
is appreciated.

Signed-off-by: Cleber Rosa <crosa@redhat.com>