Python libs: enable machine type auto selection
The arm and aarch64 targets are noteworthy because they don't have a
default machine type.  This introduces and opt-in option to users of
the qemu.machine.QEMUMachine Python class, in which a default machine
type will be set when a number of conditions are met.

The motivation for this feature is that a number of (acceptance) tests
are target agnostic, and should run with all tarets (in theory while
it's pending continuous test execution).  These tests can be listed
with:

  avocado list --filter-by-tags='-arch' --filter-by-tags-include-empty \
               --filter-by-tags-include-empty-key tests/acceptance/

Combined with the "qemu_bin" parameter that acceptance tests take, one
can run those tests with arm and aarch64 targets without adding
specific logic in each test.  To execute those with a aarch64 QEMU
binary one can run:

  avocado run --filter-by-tags='-arch' --filter-by-tags-include-empty \
              --filter-by-tags-include-empty-key \
              -p qemu_bin=aarch64-softmmu/qemu-system-aarch64 \
              tests/acceptance/

When this feature is activated, the default automatic selection of
a machine type will be logged (as it will also be done in the QEMU
command line that is already logged).

Signed-off-by: Cleber Rosa <crosa@redhat.com>